remove: duplicate code
